+++ src/Pegasus/Repository/CIMRepository.cpp
@@ -1702,7 +1702,7 @@ CIMObjectPath CIMRepository::_createInstance(
     {
         Buffer data;
         streamer->encode(data, cimInstance);
-        size = data.size();
+        size = (Uint32)data.size();
 
         if (!InstanceDataFile::appendInstance(dataFilePath, data, index))
         {
@@ -2132,7 +2132,7 @@ void CIMRepository::modifyInstance(
         streamer->encode(out, cimInstance);
         //XmlWriter::appendInstanceElement(out, cimInstance);
 
-        newSize = out.size();
+        newSize = (Uint32)out.size();
 
         if (!InstanceDataFile::appendInstance(dataFilePath, out, newIndex))
         {
@@ -2284,14 +2284,14 @@ Boolean CIMRepository::_loadAllInstances(
 
         CIMInstance tmpInstance;
 
-        Uint32 bufferSize = data.size();
+        Uint32 bufferSize = (Uint32)data.size();
         char* buffer = (char*)data.getData();
 
         for (Uint32 i = 0; i < instanceNames.size(); i++)
         {
             if (!freeFlags[i])
             {
-                Uint32 pos=(&(buffer[indices[i]]))-buffer;
+                Uint32 pos= (Uint32)((&(buffer[indices[i]]))-buffer);
                 streamer->decode(data, pos, tmpInstance);
 
                 Resolver::resolveInstance (tmpInstance, _context, nameSpace,
